% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_daye.R
\name{theme_daye}
\alias{theme_daye}
\title{Daye Theme}
\usage{
theme_daye()
}
\description{
Custom notebook-styled ggplot theme
}
\details{
Custom ggplot theme to standardize plot aesthetics in the style of notebook paper
}
